Move field parser to a different file
