+++ langwatch_nlp/langwatch_nlp/studio/execute/execute_component.py
@@ -1,5 +1,6 @@
+import asyncer
 import sentry_sdk
-from langwatch_nlp.studio.parser import autoparse_fields
+from langwatch_nlp.studio.field_parser import autoparse_fields
 from langwatch_nlp.studio.parser_v2 import get_component_class, parse_component
 from langwatch_nlp.studio.utils import disable_dsp_caching, optional_langwatch_trace
 from langwatch_nlp.studio.types.events import (
@@ -37,9 +38,9 @@ async def execute_component(event: ExecuteComponentPayload):
                 trace.autotrack_dspy()
             class_name, code = parse_component(node, event.workflow)
             Module = get_component_class(component_code=code, class_name=class_name)
-            instance = Module()  # type: ignore
+            instance = Module()
             result = await asyncify(instance)(
-                **autoparse_fields(node.data.inputs or [], event.inputs)
+                **autoparse_fields(node.data.inputs or [], event.inputs)  # type: ignore
             )
 
         cost = result.cost if hasattr(result, "cost") else None
@@ -59,4 +60,4 @@ async def execute_component(event: ExecuteComponentPayload):
         raise e
     finally:
         if trace:
-            await asyncify(trace.send_spans)()
+            await asyncer.asyncify(trace.send_spans)()
